feat(multi-actions): invert actions config FK, schema + shadow write
