+++ front/lib/api/assistant/configuration.ts
@@ -38,20 +38,20 @@ import { getPublicUploadBucket } from "@app/lib/file_storage";
 import {
   AgentConfiguration,
   AgentDataSourceConfiguration,
-  AgentDustAppRunConfiguration,
   AgentGenerationConfiguration,
-  AgentRetrievalConfiguration,
   Conversation,
   DataSource,
   Mention,
   Message,
   Workspace,
 } from "@app/lib/models";
+import { AgentTablesQueryConfigurationTable } from "@app/lib/models/assistant/actions/tables_query";
 import {
+  AgentDustAppRunConfiguration,
+  AgentRetrievalConfiguration,
   AgentTablesQueryConfiguration,
-  AgentTablesQueryConfigurationTable,
-} from "@app/lib/models/assistant/actions/tables_query";
-import { AgentUserRelation } from "@app/lib/models/assistant/agent";
+  AgentUserRelation,
+} from "@app/lib/models/assistant/agent";
 import { frontSequelize } from "@app/lib/resources/storage";
 import { generateModelSId } from "@app/lib/utils";
