+++ front/lib/models/index.ts
@@ -1,22 +1,18 @@
 import { App, Clone, Dataset, Provider, Run } from "@app/lib/models/apps";
-import {
-  AgentDustAppRunAction,
-  AgentDustAppRunConfiguration,
-} from "@app/lib/models/assistant/actions/dust_app_run";
+import { AgentDustAppRunAction } from "@app/lib/models/assistant/actions/dust_app_run";
 import {
   AgentDataSourceConfiguration,
   AgentRetrievalAction,
-  AgentRetrievalConfiguration,
   RetrievalDocument,
   RetrievalDocumentChunk,
 } from "@app/lib/models/assistant/actions/retrieval";
-import {
-  AgentTablesQueryAction,
-  AgentTablesQueryConfiguration,
-} from "@app/lib/models/assistant/actions/tables_query";
+import { AgentTablesQueryAction } from "@app/lib/models/assistant/actions/tables_query";
 import {
   AgentConfiguration,
+  AgentDustAppRunConfiguration,
   AgentGenerationConfiguration,
+  AgentRetrievalConfiguration,
+  AgentTablesQueryConfiguration,
   GlobalAgentSettings,
 } from "@app/lib/models/assistant/agent";
 import {
